Move Local ServerCaps module underneath Server
